Function riot_rso_sample_app::handlers::data::champion_rotation_data

fn champion_rotation_data(
    url: &str,
    token: &str
) -> Result<ChampionRotationData, String>

Fetches the champion rotation data.

This function sends a GET request to the provided URL with the provided token as the X-Riot-Token header. If the request is successful, it deserializes the response into a ChampionRotationData struct and returns it. If the request fails, it logs the error and returns a string describing the error.

Arguments

url - A String that holds the URL for the request.

token - A String that holds the X-Riot-Token for the request.

Returns

This function returns a Result that contains a ChampionRotationData struct if the request is successful, or a String describing the error if the request fails.

Example

let url = "https://na1.api.riotgames.com/lol/platform/v3/champion-rotations";
let token = "my_token";
let champion_rotation_data = champion_rotation_data(url, token);
match champion_rotation_data {
    Ok(data) => println!("Champion rotation data: {:?}", data),
    Err(e) => println!("An error occurred: {}", e),
}