Struct riot_rso_sample_app::handlers::default::Response

pub struct Response {
    sign_in_url: String,
}

Represents the server’s response to a request with a sign-in URL.

This struct is used to generate the HTML response for the client, directing them to the sign-in page. It leverages the Askama template engine to render the default.html template with the provided sign_in_url.

Fields
sign_in_url: String

The URL to which the user should be redirected for signing in.